refactor(cli): extract auth command subsystem module (#174)
